requirements.txt
fastapi
uvicorn[standard]
pydantic
langchain
langchain_core
langchain_groq
langchain_tavily
langchain_openai
langgraph
langsmith
unsloth
transformers
torch
Pillow
groq
unstructured[all-docs]
faiss-cpu
sentence-transformers
typing_extensions
langchain-community
gradio